backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.6",
    "django-cors-headers>=4.7.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]
